+++ rollup.config.browser.js
@@ -3,10 +3,10 @@ import babel from 'rollup-plugin-babel';
 export default {
     output: {
         format: 'umd',
-        file: 'dist/li-log.browser.js'
+        file: 'dist/li-log.js'
     },
     name: 'LiLog',
-    input: 'src/browser.js',
+    input: 'src/index.js',
     plugins: [
         babel({
             exclude: ['node_modules/**']