+++ rollup.config.browser.minified.js
@@ -4,11 +4,11 @@ import minify from 'rollup-plugin-babel-minify';
 export default {
     output: {
         format: 'umd',
-        file: 'dist/li-log.browser.min.js',
+        file: 'dist/li-log.min.js',
         sourceMap: true
     },
     name: 'LiLog',
-    input: 'src/browser.js',
+    input: 'src/index.js',
     plugins: [
         babel({
             exclude: ['node_modules/**']