+++ rollup.config.js
@@ -1,0 +1,8 @@
+export default {
+    output: {
+        format: 'es',
+        file: 'index.js'
+    },
+    name: 'LiLog',
+    input: 'src/index.js'
+};