+++ src/browser-version/console.js
@@ -1,5 +1,0 @@
-import { Log } from './log.browser.js';
-
-export function toConsole(options) {
-    return new Log(options);
-}