+++ src/index.js
@@ -1,0 +1,4 @@
+import { Log as Logger } from './lib/li-log-core';
+
+export default Logger;
+