+++ src/lib/utils.js
@@ -49,6 +49,12 @@ function copyDeep(baseObj) {
     return newObj;
 }
 
+const isBrowser = (new Function("try {return this===window;}catch(e){ return false;}"))(),
+    isNode = (new Function("try {return this===global;}catch(e){return false;}"))();
+
+
 export {
-    mergeOptions
+    mergeOptions,
+    isBrowser,
+    isNode
 }